refactor(models): restructure data models and remove unused files
