--- online_quiz/lib/models/notification_item.dart
@@ -1,19 +1,0 @@
-class NotificationItem {
-  final String id;
-  final String title;
-  final String message;
-  final DateTime timestamp;
-  final String type; // 'quiz', 'result', 'announcement', 'reminder'
-  final bool isRead;
-  final String? courseId;
-
-  NotificationItem({
-    required this.id,
-    required this.title,
-    required this.message,
-    required this.timestamp,
-    required this.type,
-    required this.isRead,
-    this.courseId,
-  });
-}
